[keymap(kyria)] moved OLED & encoder implementation to separate classes
